I finally start to understand the lance
Unfortunatelly I spend a lot of time and efort in understanding the
architecture of lance. The online game are not easy to make there a lot
to know and understand, I am glad lance cover a lot of net communication
ohtervise this project will be impossible.

I have added pug as my template engine, now I am straggling to make it
work poperly. The good thing will be add vue to this project put as I
use aframe that use dom elements. I hope with vue it will be easiesr to
add new element to the scene.