Fix the wrong include order
The was a bug that saied that some "automatic" property of the
gameEngine was not defined. It's not possible as all this staff are
defined by the framework itself.
Indeed the problem was due to the wrong order of the inclusion. The
bundel must go first before the aframe tags, otherwise it will be
difficult to add this elements to the DOM.

I realy glad I manage to overcame this issue.